+++ src/Datasets/Labeled.php
@@ -262,15 +262,15 @@ class Labeled extends Dataset
                 $desc['skewness'] = Stats::skewness($this->labels, $mean);
                 $desc['kurtosis'] = Stats::kurtosis($this->labels, $mean);
 
-                $percentiles = Stats::percentiles($this->labels, [
-                    0, 25, 50, 75, 100,
+                $quartiles = Stats::quantiles($this->labels, [
+                    0.0, 0.25, 0.5, 0.75, 1.0,
                 ]);
 
-                $desc['min'] = $percentiles[0];
-                $desc['25%'] = $percentiles[1];
-                $desc['median'] = $percentiles[2];
-                $desc['75%'] = $percentiles[3];
-                $desc['max'] = $percentiles[4];
+                $desc['min'] = $quartiles[0];
+                $desc['25%'] = $quartiles[1];
+                $desc['median'] = $quartiles[2];
+                $desc['75%'] = $quartiles[3];
+                $desc['max'] = $quartiles[4];
 
                 break 1;