+++ tests/AnomalyDetectors/LodaTest.php
@@ -155,7 +155,7 @@ class LodaTest extends TestCase
         $training = $this->generator->generate(self::TRAIN_SIZE);
         $testing = $this->generator->generate(self::TEST_SIZE);
 
-        $folds = $training->fold(3);
+        $folds = $training->stratifiedFold(3);
 
         $this->estimator->train($folds[0]);
         $this->estimator->partial($folds[1]);